cleaned up role
